menu/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-modeltranslation==0.19.5
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
